pyproject.toml
[project]
name = "aicc"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "celery>=5.5.3",
    "fastapi>=0.121.1",
    "faster-whisper>=1.2.1",
    "httpx>=0.28.1",
    "huggingface-hub>=0.20.0,<1.0.0",
    "librosa>=0.11.0",
    "loguru>=0.7.3",
    "numpy>=2.3.0",
    "psutil>=7.1.3",
    "pyannote-audio>=4.0.1",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "redis>=7.0.1",
    "soundfile>=0.13.1",
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "uvicorn[standard]>=0.38.0",
    "watchdog>=6.0.0",
]
